simulator/requirements.txt
pymodbus
PyYAML 